Separate bot from backend
- in the future backend will be a separate backend app. This commit is one step toward it
