+++ good_advice_bot/__init__.py
@@ -1,0 +1,3 @@
+from dotenv import load_dotenv
+
+load_dotenv()
