+++ good_advice_bot/__main__.py
@@ -1,0 +1,24 @@
+import asyncio
+import logging
+import os
+
+from good_advice_bot.bot.main import init_bot
+
+logging.basicConfig(
+    format="%(asctime)s:%(levelname)s:%(name)s:%(lineno)s:%(message)s",
+    level="DEBUG",
+)
+logging.getLogger("asyncio").setLevel(logging.WARNING)
+
+logger = logging.getLogger(__name__)
+
+BOT_TOKEN = os.environ["BOT_TOKEN"]
+
+
+async def main() -> None:
+    bot = init_bot(BOT_TOKEN)
+    await bot.start_polling()
+
+
+if __name__ == "__main__":
+    asyncio.run(main())
