+++ pyproject.toml
@@ -1,6 +1,7 @@
 [tool.isort]
 line_length = 100
 multi_line_output = "3"
+known_first_party = "good_advice_bot"
 include_trailing_comma = "true"
 sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
 default_section = "THIRDPARTY"
@@ -25,9 +26,16 @@ disallow_incomplete_defs = true
 disallow_untyped_calls = true
 warn_unused_ignores = true
 show_error_codes = true
+disable_error_code = [
+    "method-assign"  # when enabled, does not allow to assign to a method of a class
+]
 
 [[tool.mypy.overrides]]
 module = "asyncpg"
+ignore_missing_imports = true
+
+[[tool.mypy.overrides]]
+module = "asyncpg.protocol.protocol"
 ignore_missing_imports = true
 
 [tool.ruff]
